Tighten up tests for FrameOptions
The unit tests for the FrameOptions class should use assertEquals, not
assertContains, as the test is suppose to be checking the entirety of
the returned string, rather than a portion of it.

These tests were based off the tests for the ContentSecurityPolicy
class, which uses assertContains given how long and complex the
generated header can be.