Rebuild composer.lock for PHP 7.0 on Travis CI
The first Travis CI build failed due to the composer.lock file having
been generated on PHP 7.1. This caused doctrine/instantiator 1.1.0 to
fail to be installed, which depends on PHP 7.1.

This lock file was generated using PHP 7.0, and now the instantiator
package is 1.0.5, which is compatible with PHP 7.0.

It's possible the composer.lock file shouldn't be included so that
Composer can pick the most appropriate version of each package, however
that seems to go against most advice to always check-in the lock file.