Require PHP 7.1+
This commit bumps the minimum version to PHP 7.1 as 7.0 is unsupported
and is holding back better development. By moving to 7.1, a newer
version of PHPStan was able to pick up more errors (by way of being
able to read the `phpstan.neon' file) and Infection can now be used.

Given PHP 7.0 -should- be unused (it's an unsupported version), this
is only a minor version bump.